Actor generalization and GameMaster refactor
This version generalizes the Player class into an abstraction that can be extended by Ghosts later.   The Player class is then refactored into a PacMan concrete implementation of it, gaining player-interface functionality via the PlayerInput component.  Finally, I refactored the GameMaster to Gameboard and restructured the Unity hierarchy to a more sensible setup.