Basic Strategy Pattern and a few bug fixes
Added a rough implementation of a simple strategy pattern that allows the ghosts to use their own unique personalities to try to capture PacMan.  I also fixed a bug that allowed the player to clip into a wall if the keyboard input did not match the player's current facing at the moment a collision is detected (i.e. Holding down while PacMan moves left or right will eventually clip into a wall).