Power Pellet Activation
Update includes all the functionality expected for the activation of a power pellet -- ghosts become frightened and collision with pacman results in becoming eaten and awards points to the player based on the number of ghosts eaten for that phase.  Ghosts will retreat to the ghost house to be respawned before continuing their pursuit of PacMan.